Added Lockers, Random Key Generator, and Locker Rental Function.
Added in a nested dictionary called lockers that provides access to each individual locker. For each locker, they have the attributes Available, Key, and Items. Available determines if the locker is available for rental or not. Key holds the key value of the locker. Items are the items that are stored in the locker. The generate_key function randomly generates a key of 6 integer numbers for the locker being rented. The rent_locker() function looks for an available locker, rents it out, and gives the customer their pin for the locker. Added in placeholders for place_items() and locker_return() functions. Going to work on these functions next and continue testing them before adding in the user interaction in the main section.