Added Locker Access Function
I created a function called access_locker() that asks the customer for their locker number and the key / pin code to open their locker. If the key the customer gives is correct -- they gain access to their locker. If it is not correct - they don't get access to their locker. It is a very basic set up at the moment and more is to come.