Added in Logic for Item Placement in Locker
The place_items() function asks the user if they would like to place an item in their locker. If yes, they are asked to type the name of the item. The item is then stored in their locker and they are asked again if they would like to place an item. If no, it exits. If the answer is not y, Y, n, or N, it asks for the user to enter their answer again. This is very simple logic at the moment -- I'm looking to add more complexity soon -- such as making sure access was granted to the user first before granting access to place items in locker and adding several items in the locker at once instead of asking for one item at a time.