Added Item Removal Logic for Lockers
Now the user can remove an item from their locker. The function remove_items() lets the user remove one item at a time from their locker by asking the user for the name of the item and searching their locker items for a matching item. I also made part of the code a little easier to read by switching the conditional statements for checking for Y/y or N/n for user input in place-items and remove-items by using .lower() for the user answer instead.