Added Locker Return Function.
I added in logic for the locker return function. It asks the user if they would like to return their locker - type y or n. If y - the items from the locker are removed, available is set to true, and the key is set back to None. There are a few test statements to show that these values did change before and after this implementation. If n - the function exits. If the user types something other than y or n, they are asked to try again.