Added in Menu Selection and Locker Rental from Menu Options.
In the main function, I added in a menu selection for the locker rental system for a user to interact with. There are 6 options listed in the menu. These are rent a locker, access a locker, place an item in locker, remove an item in locker, return your locker, and exit the menu. Currently, only renting a locker works as an option. The other options are under development for use within the menu options.