Added Option 5 - Return your locker - Implementation for User Integration and Made Some Minor Adjustments to Option 1 - Rent a locker
For option 5 - return your locker - the function call is now active in the user menu. It prompts the user for their locker number and key code. It checks that the locker number is valid and is currently being rented. It also checks that the key is correct before granting access to return the locker. Later, I would like to add more security such as number of attempts allowed before being locked out of the locker and locker recovery efforts. For now, it just is simple logic for returning a rented locker. As for the tweaks made for option 1 - renting a locker - I added a few checks suchs as checking if the locker number exists first before doing anything else, and then handling if the locker is available or if is not available. Additionally, giving the customer the option to back out of both renting and returning a locker in case of accidentally pushing the wrong button in the menu selection.